(telemetry) replace default influx cluster + fallback auth
